requirements.txt
amqp==2.6.1
attrs==20.3.0
billiard==3.6.3.0
celery==4.4.7
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.17
django-extensions==3.1.0
djangorestframework==3.11.2
djangorestframework-simplejwt==4.3.0
drf-yasg==1.17.1
drf-yasg2==1.19.4
factory-boy==3.1.0
Faker==5.0.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==2.1.1
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.3
kombu==4.6.11
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.9
pathlib2==2.3.5
pluggy==0.13.1
py==1.10.0
pydotplus==2.0.2
PyJWT==1.7.1
pyparsing==2.4.7
pytest==6.1.2
pytest-factoryboy==2.0.3
python-dateutil==2.8.1
python-magic==0.4.20
pytz==2020.5
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.3
vine==1.3.0
websocket-client==0.57.0
zipp==1.2.0
